start separating out ui segment
